+++ CHANGELOG.md
@@ -7,11 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Repository: Rework `hack` & `config`. ([#176](https://github.com/giantswarm/cluster-api-app/pull/176))
+
 ## [1.15.1] - 2024-01-15
 
 ### Changed
 
-- Make: Use `kubectl` integrated `kustomize`
+- Make: Use `kubectl` integrated `kustomize`. ([#171](https://github.com/giantswarm/cluster-api-app/pull/171))
 - Switch to image hosted on gsoci.azurecr.io
 
 ## [1.15.0] - 2023-12-20